fix(dns): snapshot DNS entries once per cleanup, dedup test mocks
Per CodeRabbit review: unregisterIfOwned called dnsServer.listEntries()
per hostname (up to 4x per cleanup), each acquiring the lock and copying
the whole entries map. Snapshot once per unregisterAllAliases call --
each hostname is an independent key, so this can't go stale mid-call.

Also extracts AutoRemoveMock into the shared StaticSnapshotClientMock,
which CodeRabbit found byte-for-byte identical to a mock added later on
the stacked fix/stop-dns-cleanup branch.
