+++ Tests/socktainerTests/Routes/ContainerStartRouteAutoRemoveAliasTests.swift
@@ -21,7 +21,7 @@ struct ContainerStartRouteAutoRemoveAliasTests {
         let network = "myapp_default"
 
         let snapshot = try makeSnapshot(nativeId: nativeId, ip: ip, network: network)
-        let mock = AutoRemoveMock(snapshot: snapshot)
+        let mock = StaticSnapshotClientMock(snapshot: snapshot)
         let dnsServer = SocktainerDNSServer()
 
         // Marked --rm at create time; ContainerStartRoute doesn't touch this cache entry itself.
@@ -66,29 +66,4 @@ private func makeSnapshot(nativeId: String, ip: String, network: String) throws 
             "com.docker.compose.project": "myapp",
         ]
     )
-}
-
-private actor AutoRemoveMock: ClientContainerProtocol {
-    private var snapshot: ContainerSnapshot
-
-    init(snapshot: ContainerSnapshot) {
-        self.snapshot = snapshot
-    }
-
-    func list(showAll: Bool, filters: [String: [String]]) async throws -> [ContainerSnapshot] { [snapshot] }
-    func getContainer(id: String) async throws -> ContainerSnapshot? { snapshot }
-    nonisolated func enforceContainerRunning(container: ContainerSnapshot) throws {}
-    func start(id: String, detachKeys: String?) async throws {
-        await ContainerExitCodeStore.shared.remove(id: id)
-    }
-    func stop(id: String, signal: String?, timeout: Int?) async throws {}
-    func restart(id: String, signal: String?, timeout: Int?) async throws {}
-    func kill(id: String, signal: String?) async throws {}
-    func delete(id: String) async throws {}
-    func wait(id: String, condition: ContainerWaitCondition) async throws -> RESTContainerWait {
-        RESTContainerWait(statusCode: 0)
-    }
-    func prune(filters: [String: [String]]) async throws -> (deletedContainers: [String], spaceReclaimed: Int64) {
-        ([], 0)
-    }
 }
